Fix failure to get source maps on windows with base change
